pyproject.toml
[project]
name = "telegram-currency-bot"
version = "0.1.2"
description = "Telegram currency bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "aiosqlite>=0.21.0",
    "asyncio>=3.4.3",
    "python-dotenv>=1.0.0",
    "playwright>=1.51.0",
    "pytelegrambotapi>=4.26.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.40"
]
